map_etl/pyproject.toml
[tool.poetry]
name = "dxmaps_etl"
version = "0.1.0"
description = ""
authors = ["Carlos Arroyo <dimago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
environs = "^8.0.0"
click = "^7.1.2"
logzero = "^1.5.0"
requests = "^2.24.0"
toolz = "^0.10.0"
geojson = "^2.5.0"
googlespider = "^0.1.7"
google-api-client = "^3.14.159265359"
python-slugify = "^4.0.1"
dateparser = "^0.7.6"
pillow = "^7.2.0"
imagehash = "^4.1.0"
opencv-contrib-python = "^4.4.0"
shapely = "^1.7.1"
scikit-learn = "^0.23.2"
pydash = "^4.8.0"
pymongo = "^3.11.0"
pyquery = "^1.4.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
